ITM-842: Break up server code into generic and domain-specific components (#119)
1. Break up server code into generic and domain-specific components
2. Added domain to start_session external API
3. Added documentation. Updated swagger.yaml to remove domain-specific content from non-domain-specific areas.
