+++ swagger_server/config_util.py
@@ -26,7 +26,8 @@ def check_ini():
                 "EVAL_NAME", "EVAL_NUMBER", "SOARTECH_EVAL_QOL_SCENARIOS", "SOARTECH_EVAL_VOL_SCENARIOS", "SOARTECH_TRAIN_QOL_SCENARIOS", "SOARTECH_TRAIN_VOL_SCENARIOS",
                 "SOARTECH_QOL_ALIGNMENT_TARGETS", "SOARTECH_VOL_ALIGNMENT_TARGETS", "SOARTECH_EVAL_FILENAMES", "SOARTECH_TRAIN_FILENAMES",
                 "ADEPT_EVAL_MJ_SCENARIOS", "ADEPT_EVAL_IO_SCENARIOS", "ADEPT_TRAIN_MJ_SCENARIOS", "ADEPT_TRAIN_IO_SCENARIOS",
-                "ADEPT_MJ_ALIGNMENT_TARGETS", "ADEPT_IO_ALIGNMENT_TARGETS", "ADEPT_EVAL_FILENAMES", "ADEPT_TRAIN_FILENAMES")
+                "ADEPT_MJ_ALIGNMENT_TARGETS", "ADEPT_IO_ALIGNMENT_TARGETS", "ADEPT_EVAL_FILENAMES", "ADEPT_TRAIN_FILENAMES",
+                "DEFAULT_DOMAIN", "SUPPORTED_DOMAINS")
         for option in keys:
             if option in ini["DEFAULT"]: # checks if "option" key exists in "default" section
                 key_value = ini.get("DEFAULT", option)
