+++ swagger_server/itm/__init__.py
@@ -1,1 +1,5 @@
 from .itm_session import ITMSession
+from .itm_action_handler import ITMActionHandler
+from .itm_scenario import ITMScenario
+from .itm_scene import ITMScene
+from .itm_scenario_reader import ITMScenarioReader
