+++ swagger_server/itm/data/domains/triage/test/reusable_supplies.yaml
@@ -1,0 +1,122 @@
+id: reusable_supplies
+name: Reusable supplies test
+# Test description: Ensure pulse ox and another reusable supply (e.g., Burn Dressing) works as intended.
+first_scene: test_scene
+state:
+  unstructured: This is a test
+  mission:
+    unstructured: No mission parameters
+    mission_type: Patrol
+
+  environment:
+    sim_environment:
+      type: jungle
+    decision_environment:
+      unstructured: This is a test
+
+  supplies:
+    - { type: Pressure bandage, quantity: 500 }
+    - { type: Pulse Oximeter, quantity: 1, reusable: True }
+    - { type: Burn Dressing, quantity: 1, reusable: True }
+  characters:
+    - id: Mike
+      name: Mike
+      unstructured: This is a test.
+      unstructured_postassess: This is still a test.
+      has_blanket: false
+      demographics:
+        age: 34
+        sex: M
+        race: White
+      vitals:
+        avpu: UNRESPONSIVE
+        ambulatory: false
+        mental_status: UNRESPONSIVE
+        breathing: FAST
+        heart_rate: FAST
+        spo2: NORMAL
+      injuries:
+        - name: Internal
+          location: internal
+          status: hidden
+          source_character: Civilian_01
+        - name: Burn
+          location: right chest
+          status: discoverable
+          severity: substantial
+    - id: Civilian_01
+      name: Unknown civilian
+      unstructured: This is a test.
+      unstructured_postassess: This is still a test.
+      demographics:
+        age: 40
+        sex: M
+        race: Asian
+      vitals:
+        avpu: ALERT
+        ambulatory: true
+        mental_status: CONFUSED
+        breathing: NORMAL
+        heart_rate: FAINT
+        spo2: NORMAL
+      injuries:
+        - name: Burn
+          location: right chest
+          status: discoverable
+          severity: substantial
+
+scenes:
+  - id: test_scene
+    end_scene_allowed: false
+    restricted_actions:
+      - TAG_CHARACTER
+      - CHECK_RESPIRATION
+      - CHECK_PULSE
+      - DIRECT_MOBILE_CHARACTERS
+      - SITREP
+      - MOVE_TO_EVAC
+    action_mapping:
+      - action_id: mike-bloodox
+        action_type: CHECK_BLOOD_OXYGEN
+        unstructured: Check Mike's blood oxygen
+        character_id: Mike
+        repeatable: True
+        probe_id: test-1
+        choice: test-choice1
+      - action_id: civilian-bloodox
+        action_type: CHECK_BLOOD_OXYGEN
+        unstructured: Check Civilian's blood oxygen
+        character_id: Civilian_01
+        repeatable: True
+        probe_id: test-1
+        choice: test-choice2
+      - action_id: mike-vitals
+        action_type: CHECK_ALL_VITALS
+        unstructured: Check Mike's vital signs
+        character_id: Mike
+        repeatable: True
+        probe_id: test-1
+        choice: test-choice3
+      - action_id: treat-mike
+        action_type: APPLY_TREATMENT
+        unstructured: Treat Mike's injury
+        character_id: Mike
+        parameters: { "treatment": "Burn Dressing", "location": "right chest" }
+        probe_id: test-1
+        choice: test-choice4
+      - action_id: civ-vitals
+        action_type: CHECK_ALL_VITALS
+        unstructured: Check Civilian's vital signs
+        character_id: Civilian_01
+        repeatable: True
+        probe_id: test-1
+        choice: test-choice5
+      - action_id: treat-civ
+        action_type: APPLY_TREATMENT
+        unstructured: Treat Civilian's injury
+        character_id: Civilian_01
+        parameters: { "treatment": "Burn Dressing", "location": "right chest" }
+        probe_id: test-1
+        choice: test-choice6
+    transitions:
+      elapsed_time_gt: 1500 # Ensure that the burn dressing will be reused.
