+++ swagger_server/itm/data/domains/triage/triageActionTimes.json
@@ -1,0 +1,28 @@
+{   
+    "treatmentTimes": {
+        "Tourniquet": 60,
+        "Pressure bandage": 120,
+        "Hemostatic gauze": 30,
+        "Decompression Needle": 60,
+        "Nasopharyngeal airway": 30,
+        "Pulse Oximeter": 20,
+        "Blanket": 10,
+        "Epi Pen": 10,
+        "Vented Chest Seal": 100,
+        "Pain Medications": 10,
+        "Fentanyl Lollipop": 10,
+        "Splint": 80,
+        "Blood": 30,
+        "IV Bag": 30,
+        "Burn Dressing": 40,
+        "ALREADY_TREATED": 10
+    },
+    "TAG_CHARACTER": 10,
+    "CHECK_ALL_VITALS": 30,
+    "CHECK_BLOOD_OXYGEN": 15,
+    "CHECK_PULSE": 10,
+    "CHECK_RESPIRATION": 10,
+    "DIRECT_MOBILE_CHARACTERS": 10,
+    "MOVE_TO_EVAC": 5,
+    "SITREP": 10
+}
